+++ client/package.json
@@ -13,6 +13,7 @@
 		"classnames": "^2.5.1",
 		"next": "14.2.4",
 		"react": "^18",
+		"react-beautiful-dnd": "^13.1.1",
 		"react-dom": "^18",
 		"reactstrap": "^9.2.2",
 		"styled-components": "^6.1.11"
