+++ client/src/components/Header/logo.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import Image from "next/image";
 import styles from "./styles.module.scss";
 import logoLight from "@assets/logo-light.svg";
