+++ client/src/components/Header/styles.module.scss
@@ -1,4 +1,31 @@
+@use "@styles/colors.module.scss";
+
+.headerRoot {
+	display: flex;
+	height: 100px;
+	width: 100vw;
+	justify-content: space-between;
+}
+
 .logo {
 	padding-top: 30px;
 	padding-left: 35px;
 }
+
+.addNewTaskBtn {
+	background-color: var(--main-purple);
+	border-radius: 2rem;
+	font-family: var(--font-regular);
+	font-size: var(--font-size-small);
+	height: var(--height-button);
+	color: colors.$white;
+	padding: 1rem 2rem;
+	border: none;
+	font-weight: bold;
+	cursor: pointer;
+	float: right;
+}
+
+// .addNewTaskBtn:hover {
+// 	background-color: var(--main-purple-hover);
+// }
