+++ client/src/components/Home/styles.module.scss
@@ -1,0 +1,22 @@
+@use "@styles/colors.module.scss";
+
+.board {
+	background-color: #e3ebfa;
+	height: 94vh;
+	display: flex;
+	width: 100vw;
+	align-content: center;
+	flex-wrap: wrap;
+}
+
+.empty {
+	color: #8895a8;
+	font-style: normal;
+	font-size: 18px;
+	line-height: 23px;
+	flex: 1;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	text-align: center;
+}
