+++ client/src/components/Home/theme.jsx
@@ -1,4 +1,3 @@
-import { useState } from "react";
 import { createGlobalStyle } from "styled-components";
 
 export const lightThemeSideBar = {
