+++ client/src/components/Sidebar/styles.module.scss
@@ -4,7 +4,7 @@
 	justify-content: space-between;
 
 	.allBoards {
-		color: #828fa3;
+		color: colors.$grey;
 		padding-left: 16px;
 	}
 
@@ -17,12 +17,12 @@
 		display: flex;
 		flex-direction: column;
 		padding-left: 16px;
-		color: #828fa3;
+		color: colors.$grey;
 	}
 
 	.sidebarLink {
 		text-decoration: none;
-		color: var(--medium-grey);
+		color: colors.$grey;
 		padding: 16px 0;
 
 		&:hover {
@@ -42,8 +42,8 @@
 	}
 
 	&.isDisplayed {
-		width: 400px;
-		height: 100vh;
+		width: 300px;
+		height: 90vh;
 		text-align: left;
 		display: flex;
 		flex-direction: column;
@@ -111,5 +111,66 @@
 
 	.darkModeIcon {
 		margin-left: 20px;
+	}
+}
+
+.sideBarBtnContainer {
+	display: flex;
+	white-space: nowrap;
+	max-height: 80px;
+	margin-top: 90px;
+}
+
+.sideBarBtn {
+	border: none;
+
+	.hideSpan {
+		padding-left: 10px;
+		margin-right: 155px;
+		height: 50px;
+		overflow: hidden;
+		max-height: 50px;
+	}
+
+	&.isShowed {
+		border: none;
+		padding: 0 30px 20px 0;
+		font-weight: 700;
+		font-size: 16px;
+		color: gray;
+		column-gap: 10px;
+		border: none;
+		border-radius: 0 45px 45px 0;
+		transition: 200ms ease-in;
+		background-color: transparent;
+		margin-bottom: 20px;
+		margin-top: 10px;
+
+		.img {
+			top: 6px;
+			margin-left: 16px;
+			padding-left: 30px;
+		}
+
+		&:hover {
+			background-color: #e4ebfa;
+			color: #635fc7;
+			left: 0;
+			padding-top: 10px;
+			max-height: 80px;
+		}
+	}
+
+	&.isHidden {
+		border: none;
+		border-radius: 0 45px 45px 0;
+		padding: 10px 20px;
+		background-color: #635fc7;
+		position: absolute;
+		bottom: 20px;
+
+		&:hover {
+			background-color: #a8a4ff;
+		}
 	}
 }
