--- client/src/components/core/SidebarButtons/styles.module.scss
@@ -1,60 +1,0 @@
-.sideBarBtnContainer {
-	display: flex;
-	white-space: nowrap;
-	max-height: 80px;
-	margin-top: 90px;
-}
-
-.sideBarBtn {
-	border: none;
-
-	.hideSpan {
-		padding-left: 10px;
-		margin-right: 155px;
-		height: 50px;
-		overflow: hidden;
-		max-height: 50px;
-	}
-
-	&.isShowed {
-		border: none;
-		padding: 0 30px 20px 0;
-		font-weight: 700;
-		font-size: 16px;
-		color: gray;
-		column-gap: 10px;
-		border: none;
-		border-radius: 0 45px 45px 0;
-		transition: 200ms ease-in;
-		background-color: transparent;
-		margin-bottom: 20px;
-		margin-top: 10px;
-
-		.img {
-			top: 6px;
-			margin-left: 16px;
-			padding-left: 30px;
-		}
-
-		&:hover {
-			background-color: #e4ebfa;
-			color: #635fc7;
-			left: 0;
-			padding-top: 10px;
-			max-height: 80px;
-		}
-	}
-
-	&.isHidden {
-		border: none;
-		border-radius: 0 45px 45px 0;
-		padding: 10px 20px;
-		background-color: #635fc7;
-		position: absolute;
-		bottom: 20px;
-
-		&:hover {
-			background-color: #a8a4ff;
-		}
-	}
-}
