+++ client/src/components/core/index.js
@@ -1,2 +1,1 @@
 export * from "./Button";
-export * from "./SidebarButtons";
