+++ client/src/layout/index.jsx
@@ -1,18 +1,16 @@
-/*
- * The layout component to be used for the app would go here
- * This would house the sidebar
- * The header a
- * and render children of this components
- */
 import Sidebar from "@components/Sidebar";
-import AddNewTaskBtn from "@components/core/AddNewTaskBtn";
+import { Header } from "@components/Header/header";
+import { MainPage } from "@components/Home/mainPage";
 
 const AppLayout = () => {
 	return (
-		<>
-			<AddNewTaskBtn />
-			<Sidebar />
-		</>
+		<div className="root">
+			<Header />
+			<div className="main-body">
+				<Sidebar />
+				<MainPage />
+			</div>
+		</div>
 	);
 };
 export default AppLayout;
