+++ client/src/pages/index.js
@@ -1,12 +1,12 @@
 import { SEO } from "@src/components/SEO";
-import HomepageContent from "@components/Home";
-import 'bootstrap/dist/css/bootstrap.min.css';
+import AppLayout from "../layout/index.jsx";
+import "bootstrap/dist/css/bootstrap.min.css";
 
 export default function Home() {
 	return (
 		<>
 			<SEO title={"Homepage"} description={"Kanban task app"} />
-			<HomepageContent />
+			<AppLayout />
 		</>
 	);
 }
