+++ client/src/styles/colors.module.scss
@@ -1,2 +1,3 @@
 $main-purple: #635fc7;
 $white: #ffffff;
+$grey: #828fa3;
