+++ client/src/styles/globals.css
@@ -1,64 +1,70 @@
-@import url('https://fonts.googleapis.com/css2?family=Plus+Jakarta+Sans:ital,wght@0,200..800;1,200..800&display=swap');
+@import url("https://fonts.googleapis.com/css2?family=Plus+Jakarta+Sans:ital,wght@0,200..800;1,200..800&display=swap");
 :root {
-  --max-width: 1100px;
-  --border-radius: 12px;
-  --font-regular: "Plus Jakarta Sans", sans-serif;
+	--max-width: 1100px;
+	--border-radius: 12px;
+	--font-regular: "Plus Jakarta Sans", sans-serif;
 
-  --font-size-tiny: 12px;
-  --font-size-small: 13px;
-  --font-size-regular: 15px;
-  --font-size-medium: 18px;
-  --font-size-large: 24px;
+	--font-size-tiny: 12px;
+	--font-size-small: 13px;
+	--font-size-regular: 15px;
+	--font-size-medium: 18px;
+	--font-size-large: 24px;
 
-  --line-height-small: 15px;
-  --line-height-regular: 19px;
-  --line-height-medium: 23px;
-  --line-height-large: 30px;
-  
-  --height-button: 48px;
-  --height-input: 40px;
-  --height-header: 97px;
-  --height-card: 88px;
+	--line-height-small: 15px;
+	--line-height-regular: 19px;
+	--line-height-medium: 23px;
+	--line-height-large: 30px;
 
-  --radius-button: 24px;
+	--height-button: 48px;
+	--height-input: 40px;
+	--height-header: 97px;
+	--height-card: 88px;
 
-  --main-purple: #635FC7;
-  --main-purple-hover: #A8A4FF;
-  --secondary-purple: rgba(99, 95,199, 0.10);
-  --secondary-purple-hover: rgba(99, 95,199, 0.25);
-  --red: #EA5555;
-  --red-hover: #FF9898;
-  --black: #000112;
-  --very-dark-grey: #20212C;
-  --dark-grey: #2B2C37;
-  --dark:#3E3F4E;
-  --medium-grey: #828FA3;
-  --light:#E4EBFA;
-  --light-grey: #F4F7FD;
-  --white: #FFFFFF;
+	--radius-button: 24px;
+
+	--main-purple: #635fc7;
+	--main-purple-hover: #a8a4ff;
+	--secondary-purple: rgba(99, 95, 199, 0.1);
+	--secondary-purple-hover: rgba(99, 95, 199, 0.25);
+	--red: #ea5555;
+	--red-hover: #ff9898;
+	--black: #000112;
+	--very-dark-grey: #20212c;
+	--dark-grey: #2b2c37;
+	--dark: #3e3f4e;
+	--medium-grey: #828fa3;
+	--light: #e4ebfa;
+	--light-grey: #f4f7fd;
+	--white: #ffffff;
+}
+
+.root {
+	display: flex;
+	flex-direction: column;
+}
+
+.main-body {
+	display: flex;
 }
 
 * {
-  box-sizing: border-box;
-  padding: 0;
-  margin: 0;
+	box-sizing: border-box;
+	padding: 0;
+	margin: 0;
 }
 
 html,
 body {
-  font-family: var(--font-regular);
-  
+	font-family: var(--font-regular);
 }
 
 body {
-  color: var(--dark-grey);
-  background: var(--white);
-  font-size: var(--font-size-medium);
-  
+	color: var(--dark-grey);
+	background: var(--white);
+	font-size: var(--font-size-medium);
 }
 
 a {
-  color: inherit;
-  text-decoration: none;
+	color: inherit;
+	text-decoration: none;
 }
-
